DEV: Drop old notification id columns
The `id` column of `notifications` table and `notification_id` columns
of `shelved_notifications` and `user_badges` tables have been migrated
to bigint in previous commits (for example, 799a45a).

In order to run the migrations with zero downtime, the data had to be
copied to new columns and swapped, but the old columns have been kept
to allow for rollback.